requirements.txt
selenium
beautifulsoup4
openai
gspread
oauth2client
pandas
requests
numpy
beautifulsoup4
chromedrivermanager
webdriver-manager
openrouter
gspread
google-auth
python-docx
docx2pdf
